+++ tests/test_guiparams.py
@@ -2,17 +2,18 @@ import re
 
 import pytest
 from fmeobjects import FMEFeature
-from hypothesis import given, strategies as st, example, settings
+from hypothesis import example, given, settings
+from hypothesis import strategies as st
 
 from fmetools.guiparams import (
-    GuiType,
-    parse_gui_type,
     BoolParser,
-    IntParser,
     FloatParser,
-    StringParser,
-    ListParser,
     GuiParameterParser,
+    GuiType,
+    IntParser,
+    ListParser,
+    StringParser,
+    parse_gui_type,
 )
 
 
@@ -67,16 +68,20 @@ def test_floatparser(value):
     if value == "":
         assert parser(value) is None
         return
-    # float() can also parse "1E5", "1e0", etc.
-    if not re.match(r"^\d+(\.\d+)?\s*$", value) and not re.match(
-        r"^\d+E\d+\s*$", value, re.I
-    ):
+    valid_float = True
+    try:
+        # float() can also parse "1E5", "1e0", etc.
+        float(value)
+    except ValueError:
+        valid_float = False
+
+    if valid_float:
+        parsed = parser(value)
+        assert parsed == parser(parsed)
+        assert isinstance(parsed, float)
+    else:
         with pytest.raises(ValueError):
             parser(value)
-        return
-    parsed = parser(value)
-    assert parsed == parser(parsed)
-    assert isinstance(parsed, float)
 
 
 @given(value=st.text(), encoded=st.booleans())